Grammar fixes
Mod is great but the grammar is honestly painful. This change simply fixes all the grammar errors and typo's with it! I also improved a lot of the dialogue to fit more MC's personality, which is supposed to be a bit empty and neutral so the reader can agree with what he says.

Also fixed Monika's personality a bit, which is supposed to be more cheerful (and MC is supposed to be unconfident with her).

Please do approve of these changes, I spent quite a bit of time on them :p